[Maps] Fixes for Age options. Refactor of parseCombinedFilter
